Continued Parser (after meeting 2)
All provided basic tests pass with this commit. The newer tests are neither added or have ever been run yet. We need to standardize the error messages' text, would probably need a helper method for that. Also, the code is not super readable at the moment, and code cleanup warnings should be fixed as well.